Dodaje survey do package
