+++ .gitignore
@@ -2,4 +2,4 @@
 target/
 jjdz3-peanut-software.iml
 *.iml
-idea/workspace.xml
+.idea/workspace.xml
