+++ src/main/java/MainOptions.java
@@ -1,3 +1,5 @@
+import Survey.Option;
+
 /**
  * Created by moody on 24.02.17.
  */
