+++ src/main/java/Survey/SurveyAnswerReader.java
@@ -1,3 +1,5 @@
+package Survey;
+
 import java.util.Scanner;
 
 /**
