+++ src/main/java/Survey/SurveyLoop.java
@@ -1,3 +1,5 @@
+package Survey;
+
 /**
  * Created by moody on 24.02.17.
  */
